requirement.txt
discord
github
threading
PyGithub
wheel
python-dotenv